Bug fixes: New character now starts with basic equipment. Repair weapon spell now does something. Inventory menu won't display ATK, AP, and Wait of items, as there aren't used. Adjective assignment formula tweaked, new characters should get average ones now. inv[].type: double -> int.

+++ globals.h
@@ -151,7 +151,7 @@ struct Inventory {
     double maxcon;
     double con;
     double str;
-    double type;
+    int type;
     double part;
     double effect;
 };
